requirements.txt
langchain
langchain-huggingface
transformers
# torch
accelerate
gradio
uvicorn
fastapi
huggingface_hub
mcp
fastapi-mcp
hf_xet
langchain-ollama
ollama
langchain-mcp-adapters
langchain-core
google-genai

# vector database clients
qdrant-client
langchain-qdrant

markdownify
deepagents
tavily-python